MCU-123 LPC2148: Rename files for better conformance to naming conventions
